+++ NEWS.md
@@ -11,7 +11,7 @@
 - Now the ... arg is for `bs4SidebarMenuSubItem()`
 
 ## Minor Changes
-- Fix #41: add disable argument to `bs4DashSidebar()`. Thanks @jamiebono 
+- Fix #41: add disable argument to `bs4DashSidebar()` and `bs4DashControlbar()`. Thanks @jamiebono 
 - images are centered in `userPost()`
 - add startExpanded arg to `bs4SidebarMenuItem()`
 - add `bs4TabSetPanel()` to the gallery